load stac items incrementally
This change implements parseMetadata as an asynchronous generator, which
yields incrementally improved versions of the stac item, as soon as
necessary information is available. The last stac item yielded by the
generator is the complete item. This can be used to show a partial stac
item as soon as anything is available.